Update requirements.txt
The correct way to generate a requirements.txt file for a repository is to navigate to the repository's directory and type these commands:
pip install pipreqs
pipreqs . --force
This will search the import statements in the repository against the installed packages to generate the requirements.txt file for the repository. 
Rather than compiling a list of all your installed packages, it compiles a list of packages you are using in that repository.